Add Announcement Logic
Added the `empty style` object to the `AdminAvatarField` and use it when there is no
avatar in the component state.
Updated the `AdminRouter`. Split the Announcement and Announcement detailed pages.
Added the `AdminAddAnnouncementForm` component to adding the new announcement entity.
Refactored `AdminAnnouncemnt` container to render the new add form component.
Remove the `admin` directory in the Announcement directory.
Updated the `removeBase64` until to return the empty array when there is no
possibility to cut the prefix.
Remove test images form a static directory.