Project UI
Updated the `renderNewsNotFoundError` function in the `newsController` module.
Updated the `renderAnnouncementNotFoundError` function in the `announcementController` module.
Added the `projectController` handler to the router module.
Added the Project block to the header.pug file.
Added the Project block to the main-content.pug file.
Updated the Project template.
Added projectController, projectRenderer and projectService modules.
Update the `dependencies.json` file.