Press UI
Updated the `getDetailedProjectService` function in the `projectService` module.
Added the `pressController` handler to the router module.
Added the Press block to the header.pug file.
Added the Press block to the main-content.pug file.
Added the Press pug template.
Added pressController, pressRenderer and pressService modules.
Update the `dependencies.json` file.