Admin PressPost Section
Created the PressPost administration block.
Created the appropriate services, components and containers using the structure
of Admin NewsPost section.
Fixed the type in the `manage_apps` module.
Updated the `AdminRouter` to render the `AdminProject` container and
`AdminProjectItem` component.
Remov the `getTranslation` and `getLinks` methods from utils and in the appropriate components.