Statistic script
Created the general `app_script` module for running
the user command fro the shell.
Created the script for the getting the daily statistics
from the redis and settings it to the postgres.
Added the scripts directory for useful functionality storing.
Refactored the `clean_outdated_announcements` script
according to the new commands process.