Google My Business Service
Created the googlemybusiness application with the core Google My Business service.
Created the GoogleMyBusinessAccount model that bound the site library user and Google Business
service user's data.
Created the GoogleMyBusinessAPIService that represents the core logic for calling the Google API.
Updated the googlemybusiness view. Added the saving user's account to the GoogleMyBusinessAccount table.
Added the appropriate responses to responses.py.